pyproject.toml
# TODO: add more info, see https://github.com/pypa/sampleproject/blob/main/pyproject.toml
[project]
name = "yaroc"  # Required
version = "0.1"
authors = [
  {name = "Lukáš Poláček", email = "lukas@polacek.email" }
]
description = "Yet Another ROC"
keywords = ["orienteering", "online control"]
requires-python = ">=3.10"

dependencies = [
  "pydbus==0.6.*", # note: Linux only
  "pyudev==0.24.*", # note: Linux only
  "paho-mqtt==1.6.*",
  "requests==2.28.*",
  "protobuf==4.22.*",
  "sportident",
]

[project.optional-dependencies] # Optional
dev = [
  "black",
  "mypy",
  "isort",
  "flake8",
  "Flake8-pyproject",
  "types-paho-mqtt",
  "types-requests"
]
pylsp = [
  "python-lsp-black",
  "pylsp-mypy",
  "pyls-isort",
  "pyls-flake8",
]
test = ["pytest"]

[project.scripts]  # Optional
mqtt-bench = "yaroc.scripts.mqtt_bench:main"
mqtt-listener = "yaroc.scripts.mqtt_listener:main"
send-punch = "yaroc.scripts.send_punch:main"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
exclude = ".git|.hg|.mypy_cache|.venv|_build|build|dist|_pb2.py"

[tool.flake8]
max-line-length = 100
exclude = "**/*_pb2.py"

[tool.isort]
include_trailing_comma = true
line_length = 100
multi_line_output = 3
skip_glob = "**/*_pb2.py?"
